优化LRU缓存实现
